feat: 예약이 존재하는 시간 삭제를 방어하는 기능 구현

+++ src/main/java/roomescape/service/TimeService.java
@@ -1,12 +1,13 @@
 package roomescape.service;
 
-import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.domain.reservation.Reservation;
 import roomescape.domain.time.Time;
 import roomescape.dto.time.TimeRequest;
 import roomescape.dto.time.TimeResponse;
 import roomescape.global.exception.model.ConflictException;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.TimeRepository;
 
 // TODO: 테스트 추가
@@ -14,9 +15,11 @@ import roomescape.repository.TimeRepository;
 public class TimeService {
 
     private final TimeRepository timeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public TimeService(TimeRepository timeRepository) {
+    public TimeService(TimeRepository timeRepository, ReservationRepository reservationRepository) {
         this.timeRepository = timeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<TimeResponse> findAllTimes() {
@@ -39,6 +42,10 @@ public class TimeService {
     }
 
     public void deleteTime(Long id) {
+        List<Reservation> usingTimeReservations = reservationRepository.findByTimeId(id);
+        if (usingTimeReservations.size() > 0) {
+            throw new ConflictException(String.format("[TimeId - %d] 해당 시간에 예약이 존재하여 시간을 삭제할 수 없습니다.", id));
+        }
         timeRepository.delete(id);
     }
 }
